Yoga Pose Classifier (Image)/requirements.txt
pillow
numpy
tensorflow
streamlit